+++ WEB(FE)/package.json
@@ -11,6 +11,7 @@
     "global": "^4.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.1",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.0"
   },
